+++ strato/CHANGELOG.md
@@ -20,11 +20,15 @@ so that they could be properly moved to their respective version's subsection.
 - POST `/transaction` allows users to create contracts by providing an address through the `codePtr` field
 
 ### Changed
+- `queryStrato` is now `strato-barometer`
+- `strato-barometer` commands point to a copy of `./ethereumH` to access LevelDB data
 
 ### Fixed
 
 ### Removed
-- Removed `certInfo` flag from strato-sequencer
+- Removed 'block', 'blockGO', 'canonRedis', 'compressRoundChanges' commands from blockapps-tools
+- Removed `certInfo` flag from strato-sequencer (cert is now derived from genesis block or during sync)
+- Removed unused flags, such as `brokenRefundRenable`, `cacheTransactionResults`, `faucetEnabled`, `createTransactionResults`, `gasOn`, `splitinit`, and `useSyncMode`
 
 
 ## [11.1.0] - 3/28/2023
@@ -35,7 +39,7 @@ so that they could be properly moved to their respective version's subsection.
 - `VM_DEBUGGER=bool` flag added for connecting to the VM debugger + static analysis websocket
 - Derive service provider URLs from node's network ID for testnet and production nodes
 - Update foreign keys for `BlockApps-Mercata-Asset` + `Sale` contracts whenever there is a table expansion
-- functionality to enumerate threads and their details in `/threads` endpoint of `P2PAPI`
+- Functionality to enumerate threads and their details in `/threads` endpoint of `P2PAPI`
 - `/peers` endpoint in `P2PAPI` to list peer connections and their health
 - POST `/transaction` contract creation calls will now additionally check for address state ref table entry before resolving
 - Jenkins test to ensure slipstream post sync is consistent with boot node
@@ -65,6 +69,7 @@ so that they could be properly moved to their respective version's subsection.
 - Removed unnecessary stateDiff (and threading) in the vm-runner codebase, fixing numerous sources of persistent memory build-up.
 - Removed overcomplicated attempts at solving p2p thread issue (watchdogs, canaries, semaphore, threadmap, etc)
 - `bloc/v2.2/x509/createCert` is no more
+
 
 ## [11.0.0] - 1/22/2024